requirements.txt
Pillow==9.4.0
rembg==2.0.30
requests==2.28.2
ShopifyAPI==12.2.0
tqdm==4.64.1
typeform==1.1.0
